Pre-emptive Refactor to make claimVesting smoother
