+++ src/environment.py
@@ -2,6 +2,8 @@
 import os
 
 from dotenv import load_dotenv
+from eth_typing.ethpm import URI
+from gnosis.eth import EthereumClient
 
 load_dotenv()
 NETWORK_STRING = os.environ.get("NETWORK", "gnosis")
@@ -15,3 +17,4 @@ else:
     NODE_URL = f"{BASE_NODE_URL}/{INFURA_KEY}"
 
 print(f"Using Base Node URL {BASE_NODE_URL}")
+CLIENT = EthereumClient(URI(NODE_URL))
