+++ src/safe.py
@@ -1,13 +1,19 @@
 """Some Safe object Helpers"""
+from __future__ import annotations
+
+import argparse
 from dataclasses import dataclass
 
 from eth_typing.encoding import HexStr
 from eth_typing.evm import ChecksumAddress
 from gnosis.eth import EthereumClient
 from gnosis.safe import Safe, SafeOperation
+from gnosis.safe.api import TransactionServiceApi
+from gnosis.safe.multi_send import MultiSendTx
 from web3 import Web3
 
 from src.constants import ZERO_ADDRESS
+from src.multisend import post_safe_tx, build_and_sign_multisend
 
 
 def get_safe(address: str, client: EthereumClient) -> Safe:
@@ -57,3 +63,73 @@ def encode_exec_transaction(
         ],
     )
     return data
+
+
+@dataclass
+class SafeFamily:
+    """
+    Simple data class holding a Safe and a collection of Sub Safes
+    """
+
+    parent: ChecksumAddress
+    children: list[ChecksumAddress]
+
+    @classmethod
+    def from_args(cls) -> SafeFamily:
+        """Parses Instance of class from command line arguments."""
+        parser = argparse.ArgumentParser("Safe Family Arguments")
+        parser.add_argument(
+            "--parent",
+            type=str,
+            required=True,
+            help="Master Safe Address (owner of all sub safes)",
+        )
+        parser.add_argument(
+            "--sub-safes",
+            type=str,
+            required=True,
+            help="List of Ethereum addresses corresponding to Safes owned by parent safe",
+        )
+        args = parser.parse_args()
+        return cls(
+            parent=Web3().toChecksumAddress(args.parent),
+            children=[Web3().toChecksumAddress(c) for c in args.sub_safes.split(",")],
+        )
+
+    def as_safes(self, eth_client: EthereumClient) -> tuple[Safe, list[Safe]]:
+        """Constructs/Fetches and returns Safe Objects from the instance attributes"""
+        parent = get_safe(self.parent, eth_client)
+        children = []
+        for child in self.children:
+            child_safe = get_safe(child, eth_client)
+            if not child_safe.retrieve_is_owner(parent.address):
+                print(
+                    f"{parent} is not an owner of {child_safe} - transactions to fail!"
+                )
+            children.append(child_safe)
+
+        print(f"loaded Parent {parent.address} along with {len(children)} child safes")
+        return parent, children
+
+
+def multi_exec(
+    parent: Safe,
+    client: EthereumClient,
+    signing_key: str,
+    transactions: list[MultiSendTx],
+) -> int:
+    """
+    Iteratively builds and posts a multisend transaction adding `new_owner` to each child safe.
+    Requires that `parent` is a single signer on all `children`.
+    """
+    print("building an executing a multi-exec transaction")
+
+    return post_safe_tx(
+        safe_tx=build_and_sign_multisend(
+            safe=parent,
+            transactions=transactions,
+            client=client,
+            signing_key=signing_key,
+        ),
+        tx_service=TransactionServiceApi(client.get_network()),
+    )
